# devcon-api/src/swagger/definition.json
{
  "swagger": "2.0",
  "info": {
    "title": "Devcon API",
    "description": "Devcon API exposes information about Devcon, past events, schedules and our video archive. You can find more information at https://github.com/efdevcon/api",
    "version": "0.6.2",
    "contact": {
      "email": "support@devcon.org"
    },
    "license": {
      "name": "MIT"
    },
    "externalDocs": {
      "url": "https://github.com/efdevcon/api"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/speakers": {
      "get": {
        "tags": [
          "Speakers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "size",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          },
          {
            "name": "order",
            "in": "query",
            "type": "string"
          },
          {
            "name": "event",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/speakers/{id}": {
      "get": {
        "tags": [
          "Speakers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/speakers/{id}/sessions": {
      "get": {
        "tags": [
          "Speakers"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "event",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/hooks/pretalx/schedule": {
      "post": {
        "tags": [
          "Hooks"
        ],
        "description": "",
        "parameters": [
          {
            "name": "X-Webhook-Secret",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-webhook-secret",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}